third_party/blink/web_tests/fast/storage/disallowed-storage-expected.txt

This tests that storage methods throw exceptions when storage is disabled with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.sessionStorage threw exception SecurityError: Failed to read the 'sessionStorage' property from 'Window': Access is denied for this document..
PASS window.localStorage threw exception SecurityError: Failed to read the 'localStorage' property from 'Window': Access is denied for this document..
PASS successfullyParsed is true

TEST COMPLETE
